# docs/swagger.yaml
openapi: 3.0.3
info:
  title: GPTless Invoivcing API
  version: 1.0.0
  description: |
    API for generating invoices. Get your API key [here](https://go.gptless.au/form).

servers:
  - url: https://api.example.com

tags:
  - name: Invoices

paths:
  /v1/invoices/generate:
    post:
      tags: [Invoices]
      summary: Generate a standard invoice
      description: |
        Creates a UBL XML invoice given the correct data. If given a previous invoiceID or template invoiceID, it will fill that data into the unspecified fields.
      
      parameters:
        - $ref: "#/components/parameters/APItoken"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInvoiceRequest"

      responses:
        "201":
          description: |
            **Created**
            
            Returns the new UBL XML Invoice.

          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/UBLInvoiceExample"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the forum on our website
        
        "403":
          description: |
            **Forbidden**
            
            You do not have access to this template invoice. 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            InvoiceID of the template invoice does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

  /v1/invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      description: Provides a list of all invoices.

      parameters:
        - $ref: "#/components/parameters/APItoken"
    
      responses:
        "200":
          description: |
            **OK**
            
            Returns an array of invoiceIDs.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: number
              example: [12345, 54321]

        "401":
          description: |
            **Unauthorized** 
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the forum on our website


  /v1/invoices/{invoiceID}:
    parameters:
      - $ref: "#/components/parameters/APItoken"
      - $ref: "#/components/parameters/InvoiceId"
    get:
      tags: [Invoices]
      summary: Get invoice
      description: Given a valid invoiceID, returns the invoice XML
      responses:
        "200":
          description: |
            **OK**

            Returns the UBL XML Invoice.

          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/UBLInvoiceExample"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the forum on our website
        
        "403":
          description: |
            **Forbidden**
            
            You do not have access to this invoice.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            InvoiceID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

    delete:
      tags: [Invoices]
      summary: Delete invoice
      description: Given a valid invoiceID, deletes and returns the invoice XML
      responses:
        "200":
          description: |
            **OK**
            
            Returns the deleted UBL XML Invoice.
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/UBLInvoiceExample"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the forum on our website
        
        "403":
          description: |
            **Forbidden** 
            
            You do not have access to this invoice. 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found** 
            
            InvoiceID does not exist. 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found




  /v1/auth/register:
    post:
      tags: [Auth]
      summary: create an API user
      description: |
        creates an API user and generates their API token. These paths are developer only
      
      parameters:
        - $ref: "#/components/parameters/APIdevToken"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                groupName:
                  type: string
                  example: WamWarriors
              required:
                - groupName
                

      responses:
        "201":
          description: |
            **OK** 
            
            returns an object with the new groups API token
          content:
            application/json:
              schema:
                type: object
                properties:
                  APItoken:
                    type: string
        
        "401":
          description: Unauthorized. Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the forum on our website
                
        
        "403":
          description: |
            **Forbidden**
            
            you do not have access to make these changes. These paths are developer only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content


  /v1/auth/reset:
    put:
      tags: [Auth]
      summary: resets the API token for a group
      description: |
        given a group name returns a new token for that group
      
      parameters:
        - $ref: "#/components/parameters/APIdevToken"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                groupName:
                  type: string
                  example: WamWarriors
              required:
                - groupName
                
      responses:
        "200":
          description: |
            **OK**

            Returns an object with the new API token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  APItoken:
                    type: string

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the forum on our website
                
        
        "403":
          description: |
            **Forbidden**
            
            you do not have access to make these changes. These paths are developer only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

  /v1/auth/revoke:
    delete:
      tags: [Auth]
      summary: removes the API token for a group.
      description: |
        removes the API token for a group. note the data is still attached to the group it is just no longer accessible. Can be reactivated with /v1/auth/reset
      
      parameters:
        - $ref: "#/components/parameters/APIdevToken"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                groupName:
                  type: string
                  example: WamWarriors
              required:
                - groupName
                
      responses:
        "200":
          description: |
            **OK**
            
            returns an object with the deleted API token
          content:
            application/json:
              schema:
                type: object
                properties:
                  APItoken:
                    type: string
        
        "401":
          description: Unauthorized. Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the forum on our website
                
        
        "403":
          description: |
            **Forbidden** 
            
            you do not have access to make these changes. These paths are developer only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

components:
  parameters:
    InvoiceId:
      name: invoiceID
      in: path
      required: true
      schema:
        type: string
      description: invoiceID
      
    APItoken: 
      name: APItoken
      in: header
      required: true
      description: clients API token
      schema:
        type: string

    APIdevToken: 
      name: APIdevToken
      in: header
      required: true
      description: Developer API token
      schema:
        type: string

  schemas:
    GenerateInvoiceRequest:
      type: object
      properties:

        templateInvoice:
          type: string
          description: The InvoiceID to base unfilled fields off of. can be a template or a previous document 

        InvoiceData:
          type: object
          description: Full invoice representation.
          properties:
            invoiceID:
              type: string
            supplier:
              type: object

              properties:
                name: 
                  type: string
                ABN:
                  type: string
                  example: 12345678987
                
            customer: 
              type: object
              properties:
                name: 
                  type: string
                ABN:
                  type: string
                  example: 12345678987

            issueDate:
              type: string
              format: date
            dueDate:
              type: string
              format: date
            totalAmount:
              type: number
              format: double
            currency:
              type: string
              example: AUD

            lines:
              type: array
              items:
                type: object
                properties:
                  lineId:
                    type: string
                  description:
                    type: string
                  quantity:
                    type: number
                  unitPrice:
                    type: number
                  lineTotal:
                    type: number

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string

    UBLInvoiceExample:
      example: |
        <?xml version="1.0" encoding="UTF-8"?>
        <Invoice xmlns="urn:oasis:names:specification:ubl:schema:xsd:Invoice-2"
          xmlns:cac="urn:oasis:names:specification:ubl:schema:xsd:CommonAggregateComponents-2"
          xmlns:cbc="urn:oasis:names:specification:ubl:schema:xsd:CommonBasicComponents-2">
          <cbc:ID>123</cbc:ID>
          <cbc:IssueDate>2011-09-22</cbc:IssueDate>
          <cac:InvoicePeriod>
            <cbc:StartDate>2011-08-01</cbc:StartDate>
            <cbc:EndDate>2011-08-31</cbc:EndDate>
          </cac:InvoicePeriod>
          <cac:AccountingSupplierParty>
            <cac:Party>
              <cac:PartyName>
                <cbc:Name>Custom Cotter Pins</cbc:Name>
              </cac:PartyName>
            </cac:Party>
          </cac:AccountingSupplierParty>
          <cac:AccountingCustomerParty>
            <cac:Party>
              <cac:PartyName>
                <cbc:Name>North American Veeblefetzer</cbc:Name>
              </cac:PartyName>
            </cac:Party>
          </cac:AccountingCustomerParty>
          <cac:LegalMonetaryTotal>
            <cbc:PayableAmount currencyID="CAD">100.00</cbc:PayableAmount>
          </cac:LegalMonetaryTotal>
          <cac:InvoiceLine>
            <cbc:ID>1</cbc:ID>
            <cbc:LineExtensionAmount currencyID="CAD">100.00</cbc:LineExtensionAmount>
            <cac:Item>
              <cbc:Description>Cotter pin, MIL-SPEC</cbc:Description>
            </cac:Item>
          </cac:InvoiceLine>
        </Invoice>
      additionalProperties: true